+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 23.10.4
 
 ### Breaking Changes
+- Protection against accidental downgrade of Besu version to avoid possible data corruption problems. Starting an older version of Besu with a given data directory is now only possible with the `--allow-downgrade` configuration option [6307](https://github.com/hyperledger/besu/pull/6307)
 
 ### Deprecations
 - Forest pruning (`pruning-enabled` options) is deprecated and will be removed soon. To save disk space consider switching to Bonsai data storage format [#6230](https://github.com/hyperledger/besu/pull/6230)
@@ -13,7 +14,6 @@
 - Set Ethereum Classic mainnet activation block for Spiral network upgrade [#6267](https://github.com/hyperledger/besu/pull/6267)
 - Add custom genesis file name to config overview if specified [#6297](https://github.com/hyperledger/besu/pull/6297)
 - Update Gradle plugins and replace unmaintained License Gradle Plugin with the actively maintained Gradle License Report [#6275](https://github.com/hyperledger/besu/pull/6275)
-- Protection against accidental downgrade of Besu version to avoid possible data corruption problems. Downgrade can be enabled with `--allow-downgrade` [6307](https://github.com/hyperledger/besu/pull/6307)
 
 ### Bug fixes